fix type of value property from string to number

+++ src/Model/CaptureResponse.php
@@ -69,7 +69,7 @@ class CaptureResponse
         $formattedResponse = [
             "paymentId" => $this->paymentId,
             "settleId" => $this->settleId,
-            "value" => $this->settleId ? number_format($this->value, 2, '.', '') : 0,
+            "value" => $this->settleId ? (float) number_format($this->value, 2, '.', '') : 0,
             "requestId" => $this->requestId,
         ];
